Allow async auth providers
This adds support for async auth providers. There are many use cases for this,
for example, authenticating with Kubernetes in AWS but outside of EKS, you need
to make API calls to get a frequently rotated token. This allows that token to
be looked up in a non blocking way.

This also prepares the way for better support for reloading tokens in
Kubernetes itself given that they are now frequently rotated.